+++ review/src/main/java/com/haot/review/application/service/ReviewService.java
@@ -4,113 +4,55 @@ import com.haot.review.application.dtos.req.ReviewCreateRequest;
 import com.haot.review.application.dtos.req.ReviewSearchRequest;
 import com.haot.review.application.dtos.req.ReviewUpdateRequest;
 import com.haot.review.application.dtos.res.ReviewGetResponse;
-import com.haot.review.common.exceptions.CustomReviewException;
-import com.haot.review.common.response.enums.ErrorCode;
-import com.haot.review.domain.model.Review;
-import com.haot.review.domain.model.ReviewImage;
-import com.haot.review.domain.repository.ReviewImageRepository;
-import com.haot.review.domain.repository.ReviewRepository;
-import com.haot.review.presentation.client.LodgeClient;
-import com.haot.review.presentation.dto.LodgeReadOneResponse;
-import java.util.List;
-import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
-import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-@Service
-@RequiredArgsConstructor
-public class ReviewService {
+public interface ReviewService {
 
-  private final ReviewRepository reviewRepository;
-  private final ReviewImageRepository reviewImageRepository;
-  private final MockS3Service mockS3Service;
-  private final LodgeClient lodgeClient;
+  /**
+   * 리뷰 생성
+   *
+   * @param userId  리뷰를 작성한 사용자 ID
+   * @param request 리뷰 생성 요청 정보
+   * @return 생성된 리뷰 정보
+   */
+  ReviewGetResponse createReview(String userId, ReviewCreateRequest request);
 
-  @Transactional
-  public ReviewGetResponse createReview(String userId, ReviewCreateRequest request) {
+  /**
+   * 리뷰 단건 조회
+   *
+   * @param reviewId 조회할 리뷰의 ID
+   * @return 조회된 리뷰 정보
+   */
+  ReviewGetResponse readOne(String reviewId);
 
-    // mock S3 서비스로 이미지를 업로드 및 String List 반환
-    List<String> imageUrls = mockS3Service.uploadImages(request.images());
+  /**
+   * 리뷰 검색 조회
+   *
+   * @param role    사용자 역할 (USER, HOST, ADMIN 등)
+   * @param request 리뷰 검색 조건
+   * @param pageable 페이징 정보
+   * @return 검색된 리뷰 정보 목록 (페이징)
+   */
+  Page<ReviewGetResponse> searchReview(String role, ReviewSearchRequest request, Pageable pageable);
 
-    Review review = Review.createReview(userId, request.contents(), request.lodgeId());
-    reviewRepository.save(review);
+  /**
+   * 리뷰 업데이트
+   *
+   * @param reviewId 업데이트할 리뷰의 ID
+   * @param request  업데이트 요청 정보
+   * @param userId   요청을 보낸 사용자 ID
+   * @param role     사용자 역할 USER
+   */
+  void updateReview(String reviewId, ReviewUpdateRequest request, String userId, String role);
 
-    List<ReviewImage> reviewImages = imageUrls.stream()
-        .map(url -> ReviewImage.create(review, url))
-        .toList();
-
-    review.getImages().addAll(reviewImages);
-    reviewImageRepository.saveAll(reviewImages);
-
-    return ReviewGetResponse.of(review);
-  }
-
-  @Transactional(readOnly = true)
-  public ReviewGetResponse readOne(String reviewId) {
-
-    Review review = findActiveReviewById(reviewId);
-
-    return ReviewGetResponse.of(review);
-  }
-
-  @Transactional(readOnly = true)
-  public Page<ReviewGetResponse> searchReview(String role, ReviewSearchRequest request, Pageable pageable) {
-
-    int pageSize = (pageable.getPageSize() == 30
-        || pageable.getPageSize() == 50)
-        ? pageable.getPageSize() : 10;
-
-    Pageable customPageable = PageRequest.of(
-        pageable.getPageNumber(),
-        pageSize,
-        pageable.getSort()
-    );
-
-    return reviewRepository.searchReview(role, request, customPageable).map(ReviewGetResponse::of);
-  }
-
-  @Transactional
-  public void updateReview(String reviewId, ReviewUpdateRequest request, String userId, String role) {
-
-      Review review = findActiveReviewById(reviewId);
-
-      if (!review.getUserId().equals(userId)) {
-        throw new CustomReviewException(ErrorCode.FORBIDDEN_OPERATION);
-      }
-      review.updateReview(request.contents());
-  }
-
-  @Transactional
-  public void deleteReview(String reviewId, String userId, String role) {
-
-    Review review = findActiveReviewById(reviewId);
-
-    if (!hasPermissionToDeleteReview(review, userId, role)) {
-      throw new CustomReviewException(ErrorCode.FORBIDDEN_OPERATION);
-    }
-
-    review.deleteReview(userId);
-  }
-
-  private Review findActiveReviewById(String reviewId) {
-    return reviewRepository.findByReviewIdAndIsDeletedFalse(reviewId)
-        .orElseThrow(() -> new CustomReviewException(ErrorCode.REVIEW_NOT_FOUND));
-  }
-
-  private boolean hasPermissionToDeleteReview(Review review, String userId, String role) {
-    return switch (role) {
-      case "USER" -> review.getUserId().equals(userId);
-      case "HOST" -> isLodgeHost(userId, review.getLodgeId());
-      case "ADMIN" -> true;
-      default -> false;
-    };
-  }
-
-  private boolean isLodgeHost(String hostId, String lodgeId) {
-    LodgeReadOneResponse response = lodgeClient.readOne(lodgeId).data();
-    return hostId.equals(response.lodge().hostId());
-  }
+  /**
+   * 리뷰 삭제
+   *
+   * @param reviewId 삭제할 리뷰의 ID
+   * @param userId   요청을 보낸 사용자 ID
+   * @param role     사용자 역할 (USER, HOST, ADMIN)
+   */
+  void deleteReview(String reviewId, String userId, String role);
 }
